% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_labels.R
\name{lm_labels}
\alias{lm_labels}
\title{lm_labels creates linear equation to be included in plot}
\usage{
lm_labels(x)
}
\arguments{
\item{x}{dataframe}
}
\value{
labels
}
\description{
lm_labels creates linear equation to be included in plot
}
