% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetlist.R
\name{targetlist}
\alias{targetlist}
\title{Targetlist creates dataframes from xml files. Columns: Description, Std (Calibration level), conc (concentration), analyte, rdatum (collection date) and response (analyte area/IS area).}
\usage{
targetlist(x)
}
\arguments{
\item{x}{path to xml file(s)}
}
\value{
list
}
\description{
Targetlist creates dataframes from xml files. Columns: Description, Std (Calibration level), conc (concentration), analyte, rdatum (collection date) and response (analyte area/IS area).
}
\examples{
list1 <- list.files(pattern = "*.xml")
list1 <- lapply(list1, targetlist)
}
